compares timestamps, merges biomark data into one csv, and saves it to a new file... this works - crashed once for a file naming issue - will add error checks to this

+++ analysis/empatica/avro_data_export.py
@@ -11,7 +11,7 @@ import tkinter as tk
 from tkinter import filedialog
 import pandas as pd
 
-def get_processed_biomarkers(input_directory, output_directory, subject_name, day):
+def combine_processed_biomarkers(input_directory, output_directory, subject_name, day):
     csv_data_path = os.path.join(input_directory, day, subject_name, 'digital_biomarkers','aggregated_per_minute')
     # moved files that have "_sleep-detection", "_eda", "_prv", "_step-counts" to the output directory + a new folder "processed_biomarkers"... create this folder if it doesnt exist
     biomarker_filepath = os.path.join(output_directory, subject_name,'processed_biomarkers')
@@ -21,9 +21,10 @@ def get_processed_biomarkers(input_directory, output_directory, subject_name, da
     data_bm = []
     missing_data = []
     file_order = []
+    data_name = []
 
     for file in os.listdir(csv_data_path):
-        if '_sleep-detection' in file or '_eda' in file or '_prv' in file or '_step-counts' in file:
+        if '_sleep-detection' in file or '_eda' in file or '_prv' in file or '_step-counts' in file or '_pulse-rate' in file:
             # read the first colm from each of the files and make sure they are exactly the same
             with open(os.path.join(csv_data_path, file), 'r') as f:
                 reader = csv.reader(f)
@@ -32,8 +33,9 @@ def get_processed_biomarkers(input_directory, output_directory, subject_name, da
                 file_timestamp_iso = [data[i][1] for i in range(1, len(data))]
                 file_data = [data[i][3] for i in range(1, len(data))]
                 file_missing_data = [data[i][4] for i in range(1, len(data))]
-
+                
                 file_order.append(file)
+                data_name.append(file.split('_')[2].split('.')[0])
                 timestamp_ms.append(file_timestamp)
                 timestamp_iso.append(file_timestamp_iso)
                 data_bm.append(file_data) #add label for each appended data
@@ -42,17 +44,24 @@ def get_processed_biomarkers(input_directory, output_directory, subject_name, da
     # check if the timestamp_ms are the same
     if len(set([tuple(i) for i in timestamp_ms])) == 1:
         print('Timestamps are the same')
-        
+        df = pd.DataFrame(list(zip(timestamp_ms[0], timestamp_iso[0], data_bm[0], data_bm[1], data_bm[2], data_bm[3], data_bm[4], missing_data[0])), columns =['timestamp_ms', 'timestamp_iso', data_name[0], data_name[1], data_name[2], data_name[3], data_name[4], 'missing_data'])
+        #TODO: make this ^^ changeable to any number of files/ biomarkers
+        #check if os.path.join(biomarker_filepath, 'biomarkers_combined.csv') exists, if it does, append to it
+        if os.path.exists(os.path.join(biomarker_filepath, 'biomarkers_combined.csv')):
+            df.to_csv(os.path.join(biomarker_filepath, 'biomarkers_combined.csv'), mode='a', header=False, index=False)
+        else:
+            df.to_csv(os.path.join(biomarker_filepath, 'biomarkers_combined.csv'), index=False)
     else:
-        print('ERROR: Timestamps are not the same... saving to four separate files')
+        print('ERROR: Timestamps are not the same... saving to separate files')
         for i in range(len(file_order)):
             data_name = file_order[i].split('_')[2].split('.')[0]
             df = pd.DataFrame(list(zip(timestamp_ms[i], timestamp_iso[i], data_bm[i], missing_data[i])), columns =['timestamp_ms', 'timestamp_iso', data_name, 'missing_data'])
-            df.to_csv(os.path.join(biomarker_filepath, file_order[i]), index=False)
+            if os.path.exists(biomarker_filepath, file_order[i]):
+                df.to_csv(os.path.join(biomarker_filepath, file_order[i]), mode='a', header=False, index=False)
+            else:
+                df.to_csv(os.path.join(biomarker_filepath, file_order[i]), index=False)
         return
 
-
-    return biomarker_filepath
 
     
     
@@ -77,7 +86,7 @@ for day in ['day_1', 'day_2']:
     check_this = os.listdir(os.path.join(input_directory, day))
     for subject_name in os.listdir(os.path.join(input_directory, day)):
         
-        compiled_csv_data_path = get_processed_biomarkers(input_directory, output_directory, subject_name, day)
+        combine_processed_biomarkers(input_directory, output_directory, subject_name, day)
         
         print('----------------------------------------')
         print('Finished processing subject: ', subject_name, ' for day: ', day)
